Integrate CAMDM modules and add architecture-specific encoding workflow
